+++ frontend/package.json
@@ -39,10 +39,11 @@
     "@typescript-eslint/eslint-plugin": "^8.0.0",
     "@typescript-eslint/parser": "^8.0.0",
     "@vitejs/plugin-react": "^6.0.1",
+    "@vitest/coverage-v8": "^4.1.5",
     "autoprefixer": "^10.4.19",
     "eslint": "^9.8.0",
     "eslint-plugin-react-hooks": "^5.1.0-rc.0",
-    "eslint-plugin-react-refresh": "^0.4.9",
+    "eslint-plugin-react-refresh": "^0.5.2",
     "globals": "^15.9.0",
     "jsdom": "^25.0.0",
     "postcss": "^8.4.40",
@@ -50,8 +51,7 @@
     "typescript": "^6.0.3",
     "typescript-eslint": "^8.0.0",
     "vite": "^8.0.10",
-    "vitest": "^4.1.5",
-    "@vitest/coverage-v8": "^4.1.5"
+    "vitest": "^4.1.5"
   },
   "engines": {
     "node": "^20.19.0 || >=22.12.0"
